+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changelog
 - the `set-metadata` filter now expands placeholders in the value if of type 'string'
 - the `rename` filter now allows applying regexp/group expansion to the name
 - added the `count-data` filter for counting data items
+- added `BytesSupporter` mixin
 
 
 0.0.1 (2025-10-31)